requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.43
PyYAML==6.0.2
openai==1.109.0
jinja2==3.1.6
python-multipart==0.0.20
pytest==8.4.1
httpx==0.28.1
